refactor(pirate): split payload parsing
Move Pirate Weather payload parsing into focused helper modules while preserving the client methods and helper re-exports used by tests and callers.

Tested: uv run pytest -q -n 0 --tb=short tests/test_pirate_weather_client.py tests/test_pw_coverage.py tests/test_unit_preference_bugs.py tests/test_presentation_formatters.py; uv run ruff check src tests scripts; uv run pyright

Confidence: high
